refactor: provide an interface for different perturbators
