+++ tripleagent/perturbation_agent/src/main/java/se/kth/chaos/pagent/PerturbationAgentClassTransformer.java
@@ -4,6 +4,8 @@ import jdk.internal.org.objectweb.asm.ClassReader;
 import jdk.internal.org.objectweb.asm.ClassWriter;
 import jdk.internal.org.objectweb.asm.Opcodes;
 import jdk.internal.org.objectweb.asm.tree.*;
+import se.kth.chaos.pagent.injector.IInjector;
+import se.kth.chaos.pagent.injector.impl.*;
 
 import java.io.*;
 import java.lang.instrument.ClassFileTransformer;
@@ -13,9 +15,26 @@ import java.util.Random;
 public class PerturbationAgentClassTransformer implements ClassFileTransformer {
 
     private final AgentArguments arguments;
+    private final IInjector injector;
 
     public PerturbationAgentClassTransformer(String args) {
         this.arguments = new AgentArguments(args == null ? "" : args);
+        switch (this.arguments.operationMode()) {
+            case ARRAY_ANALYSIS:
+                this.injector = new ArrayAnalysisInjectorImpl();
+                break;
+            case ARRAY_PONE:
+                this.injector = new ArrayPOneInjectorImpl();
+                break;
+            case TIMEOUT:
+                this.injector = new TimeoutInjectorImpl();
+                break;
+            case THROW_E:
+                this.injector = new ThrowExceptionInjectorImpl();
+                break;
+            default:
+                this.injector = new DefaultInjectorImpl();
+        }
     }
 
     public byte[] transform(ClassLoader loader, String className, Class<?> classBeingRedefined,
@@ -25,177 +44,14 @@ public class PerturbationAgentClassTransformer implements ClassFileTransformer {
 
     private byte[] meddle(byte[] classFileBuffer) {
         ClassReader classReader = new ClassReader(classFileBuffer);
-        ClassWriter classWriter = null;
         ClassNode classNode = new ClassNode();
 
         classReader.accept(classNode, 0);
-
-        if (inWhiteList(classNode.name)) return classFileBuffer;
-
-        switch (arguments.operationMode()) {
-            case ARRAY_ANALYSIS:
-                classNode.methods.stream()
-                        .filter(method -> !method.name.startsWith("<"))
-                        .filter(method -> arguments.filter().matches(classNode.name, method.name))
-                        .forEach(method -> {
-                            int exceptionIndexNumber = 0;
-                            InsnList insnList = method.instructions;
-                            for (AbstractInsnNode node : insnList.toArray()) {
-                                if (node.getOpcode() >= Opcodes.IALOAD && node.getOpcode() <= Opcodes.AALOAD) {
-                                    // an array reading operation
-                                    System.err.println(String.format("INFO PerturbationAgent array reading at: %s/%s", classNode.name, method.name));
-                                    AbstractInsnNode previousNode = node.getPrevious();
-                                    String readingIndex = "UNKNOWN";
-                                    if (previousNode.getOpcode() >= Opcodes.ICONST_M1 && previousNode.getOpcode() <= Opcodes.ICONST_5) {
-                                        readingIndex = previousNode.getOpcode() - 3 + "";
-                                    } else if (previousNode.getOpcode() == Opcodes.BIPUSH) {
-                                        readingIndex = ((IntInsnNode) previousNode).operand + "";
-                                    } else if (previousNode.getOpcode() == Opcodes.ILOAD) {
-                                        readingIndex = "a local variable";
-                                    }
-
-                                    System.err.println("INFO PerturbationAgent the array index is:" + readingIndex);
-
-                                    PerturbationPoint perturbationPoint = new PerturbationPoint(classNode.name, method.name, method.desc, exceptionIndexNumber,
-                                            arguments.defaultMode(), 0, arguments.chanceOfFailure(), arguments.interval());
-                                    PAgent.registerPerturbationPoint(perturbationPoint, arguments);
-                                    exceptionIndexNumber = exceptionIndexNumber + 1;
-                                }
-                            }
-                        });
-                break;
-            case ARRAY_PONE:
-                classNode.methods.stream()
-                    .filter(method -> !method.name.startsWith("<"))
-                    .filter(method -> arguments.filter().matches(classNode.name, method.name))
-                    .forEach(method -> {
-                        int indexNumber = 0;
-                        InsnList insnList = method.instructions;
-                        for (AbstractInsnNode node : insnList.toArray()) {
-                            if (node instanceof VarInsnNode && node.getOpcode() == Opcodes.ALOAD) {
-                                // an local variable array loading operation
-                                // System.out.println("INFO PerturbationAgent load an array variable");
-                            } else if (node instanceof InsnNode && node.getOpcode() >= Opcodes.IALOAD && node.getOpcode() <= Opcodes.AALOAD) {
-                                // an array reading operation
-                                // System.out.println("INFO PerturbationAgent read an array");
-                                AbstractInsnNode previousNode = node.getPrevious();
-                                String readingIndex = "UNKNOWN";
-                                if (previousNode.getOpcode() >= Opcodes.ICONST_M1 && previousNode.getOpcode() <= Opcodes.ICONST_5) {
-                                    readingIndex = previousNode.getOpcode() - 3 + "";
-                                } else if (previousNode.getOpcode() == Opcodes.BIPUSH) {
-                                    readingIndex = ((IntInsnNode) previousNode).operand + "";
-                                } else if (previousNode.getOpcode() == Opcodes.ILOAD) {
-                                    readingIndex = "a local variable, index: " + ((VarInsnNode) previousNode).var;
-                                }
-
-                                // System.out.println("INFO PerturbationAgent the array index is:" + readingIndex);
-                                PerturbationPoint perturbationPoint = new PerturbationPoint(classNode.name, method.name, method.desc, indexNumber,
-                                        arguments.defaultMode(), arguments.countdown(), arguments.chanceOfFailure(), arguments.interval());
-                                PAgent.registerPerturbationPoint(perturbationPoint, arguments);
-                                insnList.insertBefore(node, arguments.operationMode().generateByteCode(classNode, method, arguments, perturbationPoint));
-                                indexNumber = indexNumber + 1;
-                            }
-                        }
-                    });
-                break;
-            case TIMEOUT:
-                classNode.methods.stream()
-                    .filter(method -> !method.name.startsWith("<"))
-                    .filter(method -> arguments.filter().matches(classNode.name, method.name))
-                    .filter(method -> method.tryCatchBlocks.size() > 0)
-                    .forEach(method -> {
-                        int indexNumber = 0;
-                        for (TryCatchBlockNode tc : method.tryCatchBlocks) {
-                            if (tc.type.equals("null")) continue; // "synchronized" keyword or try-finally block might make the type empty
-                            if (inTimeoutExceptionList(tc.type)) {
-                                PerturbationPoint perturbationPoint = new PerturbationPoint(classNode.name, method.name, method.desc, indexNumber, tc.type, 0,
-                                        arguments.defaultMode(), arguments.countdown(), arguments.chanceOfFailure(), arguments.interval());
-                                PAgent.registerPerturbationPoint(perturbationPoint, arguments);
-                                InsnList newInstructions = arguments.operationMode().generateByteCode(classNode, method, arguments, perturbationPoint);
-                                method.instructions.insert(tc.start, newInstructions);
-                                indexNumber = indexNumber + 1;
-                            }
-                        }
-                    });
-                break;
-            case THROW_E:
-                classNode.methods.stream()
-                    .filter(method -> !method.name.startsWith("<"))
-                    .filter(method -> arguments.filter().matches(classNode.name, method.name))
-                    .filter(method -> method.exceptions.size() > 0)
-                    .forEach(method -> {
-                        int exceptionIndexNumber = 0;
-                        InsnList originalInsnList = this.copyInsn(method.instructions);
-                        for (String exception : method.exceptions) {
-                            if (arguments.exceptionFilter().matches(exception)) {
-                                switch (arguments.lineNumber()) {
-                                    case "*": {
-                                        for (int i = 0; i < originalInsnList.size(); i++) {
-                                            AbstractInsnNode currentNode = originalInsnList.get(i);
-                                            if (currentNode instanceof LineNumberNode) {
-                                                method.instructions.insertBefore(currentNode,
-                                                        this.constructPerturbationPoints(classNode, method,
-                                                                exceptionIndexNumber, exception,
-                                                                ((LineNumberNode)currentNode).line));
-                                            }
-                                        }
-                                        break;
-                                    }
-
-                                    case "0": {
-                                        AbstractInsnNode firstLineNumberNode = this.getFirstOrLastLineNumberNode(originalInsnList, false);
-                                        if (firstLineNumberNode != null) {
-                                            method.instructions.insertBefore(firstLineNumberNode,
-                                                    this.constructPerturbationPoints(classNode, method,
-                                                            exceptionIndexNumber, exception,
-                                                            ((LineNumberNode)firstLineNumberNode).line));
-                                        }
-                                        break;
-                                    }
-
-                                    case "$": {
-                                        AbstractInsnNode lastLineNumberNode = this.getFirstOrLastLineNumberNode(originalInsnList, true);
-                                        if (lastLineNumberNode != null) {
-                                            method.instructions.insertBefore(lastLineNumberNode,
-                                                    this.constructPerturbationPoints(classNode, method,
-                                                            exceptionIndexNumber, exception,
-                                                            ((LineNumberNode)lastLineNumberNode).line));
-                                        }
-                                        break;
-                                    }
-
-                                    default: {
-                                        for (int i = 0; i < originalInsnList.size(); i++) {
-                                            AbstractInsnNode currentNode = originalInsnList.get(i);
-                                            if (currentNode instanceof LineNumberNode) {
-                                                int line = ((LineNumberNode)currentNode).line;
-                                                if (arguments.lineNumber().equals(line + "")) {
-                                                    method.instructions.insertBefore(currentNode,
-                                                            this.constructPerturbationPoints(classNode, method,
-                                                                    exceptionIndexNumber, exception,
-                                                                    line));
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-
-                                exceptionIndexNumber = exceptionIndexNumber + 1;
-                            }
-                        }
-                    });
-                break;
-            default:
-                // nothing now
-                break;
+        if (inWhiteList(classNode.name)) {
+            return classFileBuffer;
+        } else {
+            return this.injector.transformClass(classFileBuffer, arguments);
         }
-
-        classWriter = new ClassWriter(ClassWriter.COMPUTE_MAXS); // COMPUTE_FRAMES does not work for some cases
-        classNode.accept(classWriter);
-
-//        writeIntoClassFile(classNode.name, classWriter.toByteArray());
-
-        return classWriter != null ? classWriter.toByteArray() : classFileBuffer;
     }
 
     private boolean inWhiteList(String className) {
@@ -210,60 +66,5 @@ public class PerturbationAgentClassTransformer implements ClassFileTransformer {
         }
 
         return result;
-    }
-
-    private boolean inTimeoutExceptionList(String className) {
-        String[] whiteList = {"java/util/concurrent/TimeoutException", "java/net/SocketTimeout"};
-        boolean result = false;
-
-        for (String prefix : whiteList) {
-            if (className.startsWith(prefix)) {
-                result = true;
-                break;
-            }
-        }
-
-        return result;
-    }
-
-    private void writeIntoClassFile(String className, byte[] data) {
-        try {
-            String[] parts = className.split("/");
-            DataOutputStream dout = new DataOutputStream(new FileOutputStream(new File(parts[parts.length - 1] + ".class")));
-            dout.write(data);
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private InsnList copyInsn(InsnList origin) {
-        InsnList result = new InsnList();
-        for (int i = 0; i < origin.size(); i++) {
-            result.add(origin.get(i));
-        }
-        return result;
-    }
-
-    private AbstractInsnNode getFirstOrLastLineNumberNode(InsnList insnList, boolean reverse) {
-        AbstractInsnNode result = null;
-        for (int i = 0; i < insnList.size(); i++) {
-            AbstractInsnNode currentNode = insnList.get(i);
-            if (currentNode instanceof LineNumberNode) {
-                result = currentNode;
-                if (!reverse) break;
-            }
-        }
-        return result;
-    }
-
-    private InsnList constructPerturbationPoints(ClassNode classNode, MethodNode method, int exceptionIndexNumber,
-            String exceptionType, int lineNumberIndex) {
-        PerturbationPoint perturbationPoint = new PerturbationPoint(classNode.name, method.name, method.desc,
-                exceptionIndexNumber, exceptionType, lineNumberIndex, arguments.defaultMode(),
-                arguments.countdown(), arguments.chanceOfFailure(), arguments.interval());
-        PAgent.registerPerturbationPoint(perturbationPoint, arguments);
-        return arguments.operationMode().generateByteCode(classNode, method, arguments, perturbationPoint);
     }
 }
