+++ tripleagent/perturbation_agent/src/main/java/se/kth/chaos/pagent/injector/IInjector.java
@@ -1,0 +1,7 @@
+package se.kth.chaos.pagent.injector;
+
+import se.kth.chaos.pagent.AgentArguments;
+
+public interface IInjector {
+    byte[] transformClass(byte[] classFileBuffer, AgentArguments arguments);
+}
