+++ tripleagent/perturbation_agent/src/main/java/se/kth/chaos/pagent/injector/impl/DefaultInjectorImpl.java
@@ -1,0 +1,11 @@
+package se.kth.chaos.pagent.injector.impl;
+
+import se.kth.chaos.pagent.AgentArguments;
+import se.kth.chaos.pagent.injector.IInjector;
+
+public class DefaultInjectorImpl implements IInjector {
+    @Override
+    public byte[] transformClass(byte[] classFileBuffer, AgentArguments arguments) {
+        return classFileBuffer;
+    }
+}
